+++ android/app/src/main/java/io/coursepick/coursepick/data/preference/DefaultPreferencesRepository.kt
@@ -1,6 +1,7 @@
 package io.coursepick.coursepick.data.preference
 
 import io.coursepick.coursepick.domain.preference.PreferencesRepository
+import io.coursepick.coursepick.domain.preference.RouteFinder
 import kotlinx.coroutines.flow.Flow
 import javax.inject.Inject
