+++ android/app/src/main/java/io/coursepick/coursepick/data/preference/PreferencesDataStore.kt
@@ -6,6 +6,7 @@ import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import io.coursepick.coursepick.di.Settings
+import io.coursepick.coursepick.domain.preference.RouteFinder
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
