+++ android/app/src/main/java/io/coursepick/coursepick/domain/preference/PreferencesRepository.kt
@@ -1,6 +1,5 @@
 package io.coursepick.coursepick.domain.preference
 
-import io.coursepick.coursepick.data.preference.RouteFinder
 import kotlinx.coroutines.flow.Flow
 
 interface PreferencesRepository {
