+++ android/app/src/main/java/io/coursepick/coursepick/presentation/preference/PreferencesViewModel.kt
@@ -3,8 +3,8 @@ package io.coursepick.coursepick.presentation.preference
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.coursepick.coursepick.data.preference.RouteFinder
 import io.coursepick.coursepick.domain.preference.PreferencesRepository
+import io.coursepick.coursepick.domain.preference.RouteFinder
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
